Separate runtime evidence record logic
Runtime artifact IO now delegates record construction and validation to a dedicated module, keeping artifact persistence focused on files and directories.

Constraint: Preserve existing runtime_artifacts imports for CLI and callers while deepening the record model.
Rejected: Splitting every artifact file into separate modules | the current size only needs one deeper record module, not a broad package rewrite.
Confidence: high
Scope-risk: moderate
Directive: Add future runtime evidence record shapes through runtime_records before wiring persistence in runtime_artifacts.
Tested: python3 -m unittest discover -s tests; python3 -m compileall src; python3 -m src.cli docs workflows --check; git diff --check HEAD
Not-tested: Backward compatibility for private _validate_delegation imports, which were intentionally not public.
